+++ totalsegmentator/statistics.py
@@ -63,7 +63,35 @@ def get_radiomics_features_for_entire_dir(ct_file:Path, mask_dir:Path, file_out:
         json.dump(stats, f, indent=4)
 
 
-def get_basic_statistics_for_entire_dir(seg: np.array, ct_file, file_out:Path, quiet:bool=False):
+# def is_in_first_or_last_slice(mask):
+#     """
+#     Check if a mask is in the first or last slice of the image.
+#     Then we do not calc any statistics for it because the mask
+#     is incomplete.
+#     """
+#     # in first or last slice segmentation of bad. Therefore they one slice before and after.
+#     first_slice = mask[:, :, 1]
+#     last_slice = mask[:, :, -2]
+#     return (first_slice.sum() > 0) or (last_slice.sum() > 0)
+
+
+def touches_border(mask):
+    """
+    Check if mask touches any of the borders. Then we do not calc any statistics for it because the mask
+    is incomplete.
+    Do not check last slice by one previous, because segmentation on last slice often bad.
+    """
+    if np.any(mask[1, :, :]) or np.any(mask[-2, :, :]):
+        return True
+    if np.any(mask[:, 1, :]) or np.any(mask[:, -2, :]):
+        return True
+    if np.any(mask[:, :, 1]) or np.any(mask[:, :, -2]):
+        return True
+    return False
+
+
+def get_basic_statistics(seg: np.array, ct_file, file_out: Path, quiet: bool = False,
+                         task: str = "total", exclude_masks_at_border: bool = True):
     """
     ct_file: path to a ct_file or a nifti file object
     """
@@ -72,14 +100,19 @@ def get_basic_statistics_for_entire_dir(seg: np.array, ct_file, file_out:Path, q
     spacing = ct_img.header.get_zooms()
     vox_vol = spacing[0] * spacing[1] * spacing[2]
     stats = {}
-    for k, mask_name in tqdm(class_map["total"].items(), disable=quiet):
+    for k, mask_name in tqdm(class_map[task].items(), disable=quiet):
         stats[mask_name] = {}
         # data = nib.load(mask).get_fdata()  # loading: 0.6s
         data = seg == k  # 0.18s
-        stats[mask_name]["volume"] = data.sum() * vox_vol  # vol in mm3; 0.2s
-        roi_mask = (data > 0).astype(np.uint8)  # 0.16s
-        # stats[mask_name]["intensity"] = ct[roi_mask > 0].mean().round(2) if roi_mask.sum() > 0 else 0.0  # 3.0s
-        stats[mask_name]["intensity"] = np.average(ct, weights=roi_mask).round(2) if roi_mask.sum() > 0 else 0.0  # 0.9s
+        if touches_border(data) and exclude_masks_at_border:
+            # print(f"WARNING: {mask_name} touches border. Skipping.")
+            stats[mask_name]["volume"] = 0.0
+            stats[mask_name]["intensity"] = 0.0
+        else:
+            stats[mask_name]["volume"] = data.sum() * vox_vol  # vol in mm3; 0.2s
+            roi_mask = (data > 0).astype(np.uint8)  # 0.16s
+            # stats[mask_name]["intensity"] = ct[roi_mask > 0].mean().round(2) if roi_mask.sum() > 0 else 0.0  # 3.0s
+            stats[mask_name]["intensity"] = np.average(ct, weights=roi_mask).round(2) if roi_mask.sum() > 0 else 0.0  # 0.9s
     
     # For nora json is good
     # For other people csv might be better -> not really because here only for one subject each -> use json